split requantification_handler.py
